Added another file for float arithmetic instructions, furthered testing all implemented instructions. Continued testing and implementation required.
